requirements.txt
flask
flask-cors
numpy
opencv-python
ultralytics
torch
torchvision
python-dotenv
supabase
deepseek
geopy